+++ packages/svelte/src/index.ts
@@ -1,4 +1,4 @@
-export * from './helpers';
-export { user, accessToken, isLoading, error } from './store';
+export { type Session, checkSession } from './helpers';
+export { isLoading, error } from './store';
 
 export { default as SupaAuthHelper } from './SupaAuthHelper.svelte';