+++ packages/sveltekit/src/app.d.ts
@@ -3,9 +3,7 @@
 // See https://kit.svelte.dev/docs/typescript
 // for information about these interfaces
 declare namespace App {
-	interface Locals {
-		user: import("@supabase/supabase-js").User;
-		accessToken: string | null;
-		error: import("@supabase/supabase-js").ApiError;
+	type AuthSession = import('./types').Locals
+	interface Locals extends AuthSession {
 	}
 }