+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @supabase/auth-helpers-svelte
 
+## 0.4.2
+
+### Patch Changes
+
+- ffbfcc8: Fix null value on first run of user subscribe
+
+## 0.4.1
+
+### Patch Changes
+
+- 2f2abf2: Fix session being overwritten in SupaAuthHelper
+
 ## 0.4.0
 
 ### Minor Changes